test: 💍 id は bigint に至るので、マイグレーションを修正した

+++ db/migrate/20210522033440_create_guilds.rb
@@ -1,7 +1,7 @@
 class CreateGuilds < ActiveRecord::Migration[6.1]
   def change
     create_table :guilds do |t|
-      t.integer :id_number
+      t.bigint :id_number
       t.string :name
       t.string :icon_url
